20AIer-xhs/xhs_auto/requirements.txt
openai>=1.0.0
pillow>=9.0.0
pyyaml>=6.0
